cleaned up business logic in domain
