+++ web-app/src/api/vite-api.ts
@@ -1,9 +1,9 @@
-import { SupportedLocale } from "@windy-civi/domain/constants";
+import { DataStoreGetter } from "@windy-civi/domain/drivers";
 import {
-  CiviLegislationData,
   CiviGptLegislationData,
-  DataStoreGetter,
-} from "@windy-civi/domain/types";
+  CiviLegislationData,
+} from "@windy-civi/domain/legislation";
+import { SupportedLocale } from "@windy-civi/domain/locales";
 
 export const viteDataGetter: DataStoreGetter = {
   getLegislationData: (locale) => legislationApi[locale](),
