+++ web-app/src/app/native-web-bridge/web-bridge.ts
@@ -1,6 +1,7 @@
-import { UserPreferences } from "../types";
-
 // Theres a monorepo issue with native apps that isn't allowing symlinked packages to work in react native.
+
+import { UserPreferences } from "@windy-civi/domain/user-preferences";
+
 // https://github.com/expo/expo/issues/9591#issuecomment-1485871356
 export const USER_PREFERENCES_CHANGED = "ON_USER_PREFERENCES_CHANGED";
 
@@ -11,7 +12,7 @@ export const publishUserPreferences = (userPreferences: UserPreferences) => {
       JSON.stringify({
         type: USER_PREFERENCES_CHANGED,
         payload: userPreferences,
-      })
+      }),
     );
   }
 };
